+++ drugs/CustomTableVIewCell.h
@@ -1,0 +1,13 @@
+//
+//  CustomTableVIewCell.h
+//  drugs
+//
+//  Created by robert on 9/26/15.
+//  Copyright (c) 2015 robert. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface CustomTableVIewCell : NSObject
+
+@end
