+++ drugs/CustomTableVIewCell.m
@@ -1,0 +1,13 @@
+//
+//  CustomTableVIewCell.m
+//  drugs
+//
+//  Created by robert on 9/26/15.
+//  Copyright (c) 2015 robert. All rights reserved.
+//
+
+#import "CustomTableVIewCell.h"
+
+@implementation CustomTableVIewCell
+
+@end
