+++ drugs/DrugViewCell.swift
@@ -1,0 +1,15 @@
+		//
+//  DrugViewCell.swift
+//  drugs
+//
+//  Created by robert on 9/26/15.
+//  Copyright (c) 2015 robert. All rights reserved.
+//
+
+import UIKit
+
+        class DrugViewCell : UITableViewCell {
+        
+            
+            
+        }
